requirements.txt
langchain
langchain-groq
langchain-google-genai
pandas
gradio
python-dotenv